Refactor auth actions and components to follow vertical slice architecture
